Removed extraneous classes
