+++ app/src/main/kotlin/demand/Amazon.kt
@@ -1,29 +1,67 @@
 package com.adsbynimbus.android.sample.demand
 
+import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
+import com.adsbynimbus.Nimbus
+import com.adsbynimbus.NimbusAdManager
 import com.adsbynimbus.NimbusError
 import com.adsbynimbus.android.sample.BuildConfig
-import com.adsbynimbus.android.sample.adManager
-import com.adsbynimbus.android.sample.request.loadAd
 import com.adsbynimbus.android.sample.databinding.LayoutInlineAdBinding
-import com.adsbynimbus.android.sample.request.DTBException
-import com.adsbynimbus.android.sample.request.loadAll
 import com.adsbynimbus.openrtb.request.Format
 import com.adsbynimbus.render.AdController
 import com.adsbynimbus.render.AdEvent
 import com.adsbynimbus.request.*
+import com.amazon.device.ads.AdError
+import com.amazon.device.ads.AdRegistration
+import com.amazon.device.ads.DTBAdCallback
+import com.amazon.device.ads.DTBAdNetwork
+import com.amazon.device.ads.DTBAdNetworkInfo
 import com.amazon.device.ads.DTBAdRequest
+import com.amazon.device.ads.DTBAdResponse
 import com.amazon.device.ads.DTBAdSize
+import com.amazon.device.ads.MRAIDPolicy
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.TimeoutCancellationException
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlinx.coroutines.withContext
+import kotlinx.coroutines.withTimeout
 import timber.log.Timber
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+
+/** Demonstrates how to initialize the APS SDK for Nimbus */
+fun Context.initializeAmazonPublisherServices(appKey: String) {
+    /* Initialize APS SDK */
+    AdRegistration.getInstance(appKey, this) // this is the application context
+
+    /* Set the MRAID Policy */
+    AdRegistration.setMRAIDSupportedVersions(arrayOf("1.0", "2.0", "3.0"))
+    AdRegistration.setMRAIDPolicy(MRAIDPolicy.CUSTOM)
+
+    /* Set Nimbus as the Mediator */
+    AdRegistration.setAdNetworkInfo(DTBAdNetworkInfo(DTBAdNetwork.NIMBUS))
+
+    /* Set Nimbus as the Open Measurement Partner */
+    AdRegistration.addCustomAttribute("omidPartnerName", Nimbus.sdkName)
+    AdRegistration.addCustomAttribute("omidPartnerVersion", Nimbus.version)
+
+    /* Optional: Enable APS logging / test mode to verify the integration */
+    AdRegistration.enableLogging(true)
+    AdRegistration.enableTesting(true)
+}
 
 class APSFragment : Fragment(), NimbusRequest.Interceptor {
 
+    val adManager: NimbusAdManager = NimbusAdManager()
     private var adController: AdController? = null
 
     override fun onCreateView(
@@ -108,4 +146,60 @@ class APSFragment : Fragment(), NimbusRequest.Interceptor {
             }
         })
     }
+}
+
+/** Wraps an [AdError] response from a [DTBAdCallback] */
+class DTBException(val error: AdError) : Exception() {
+    inline val isNoBid: Boolean get() = error.code == AdError.ErrorCode.NO_FILL
+    inline val isNetworkError: Boolean get() = error.code == AdError.ErrorCode.NETWORK_ERROR
+    inline val isRequestError: Boolean get() = error.code == AdError.ErrorCode.REQUEST_ERROR
+    inline val isUnknownError: Boolean get() = error.code == AdError.ErrorCode.INTERNAL_ERROR
+}
+
+/**
+ * A cancellable coroutine that wraps [DTBAdRequest.loadAd].
+ *
+ * @param ttl The timeout in ms the loadAd call must return by
+ * @return A bid from Amazon that can be appended directly to a NimbusRequest
+ * @throws DTBException Amazon did not bid or an error occurred during the request
+ * @throws TimeoutCancellationException Amazon did not return a response in time
+ */
+suspend fun DTBAdRequest.loadAd(ttl: Long = 750): DTBAdResponse = withTimeout(ttl) {
+    suspendCancellableCoroutine { coroutine ->
+        loadAd(object : DTBAdCallback {
+            override fun onFailure(error: AdError) {
+                coroutine.resumeWithException(DTBException(error))
+            }
+
+            override fun onSuccess(response: DTBAdResponse) {
+                coroutine.resume(response)
+            }
+        })
+    }
+}
+
+/**
+ * Loads a collection of Amazon requests in parallel and returns a list of successful responses.
+ *
+ * @param timeout the timeout in milliseconds each request must return by
+ * @param onFailedRequest an optional action called on each request failure or timeout
+ */
+suspend fun Collection<DTBAdRequest>.loadAll(
+    timeout: Long = 750,
+    onFailedRequest: (DTBAdRequest, Throwable) -> Unit = { _, _ -> },
+): List<DTBAdResponse> = coroutineScope {
+    val outerScope = coroutineContext
+    val inFlightRequests = map { dtbAdRequest ->
+        async(Dispatchers.IO) {
+            runCatching { dtbAdRequest.loadAd(ttl = timeout) }
+                .onFailure { withContext(outerScope) { onFailedRequest(dtbAdRequest, it) } }
+                .getOrNull()
+        }
+    }
+    inFlightRequests.awaitAll().filterNotNull()
+}
+
+/** Loads Amazon requests in parallel and adds the successful responses to the Nimbus request. */
+suspend inline fun NimbusRequest.addAmazonAds(vararg requests: DTBAdRequest) = apply {
+    requests.toList().loadAll().forEach { addApsResponse(it) }
 }
