+++ app/src/main/kotlin/test/TestRenderFragment.kt
@@ -6,13 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.adsbynimbus.NimbusAd
-import com.adsbynimbus.android.sample.R
+import com.adsbynimbus.NimbusAdManager
 import com.adsbynimbus.android.sample.databinding.FragmentTestRenderBinding
-import com.adsbynimbus.android.sample.unescape
 import com.adsbynimbus.render.CompanionAd
 import com.adsbynimbus.render.Renderer.Companion.loadBlockingAd
 
 class TestRenderFragment : Fragment() {
+
+    val adManager: NimbusAdManager = NimbusAdManager()
 
     private val vastRegex = Regex("<vast", RegexOption.IGNORE_CASE)
     private val htmlRegex = Regex("<html", RegexOption.IGNORE_CASE)
@@ -38,4 +39,18 @@ class TestRenderFragment : Fragment() {
             })?.start()
         }
     }.root
+
+    /** A best guess algorithm for unescaping HTML markup that may be dumped from a server log */
+    fun String.unescape(): String = replace(Regex("\\s+"), " ")
+        .replace("""\n""", " ")
+        .replace("""\u003d""", "=")
+        .replace("""\u003c""", "<")
+        .replace("""\u003e""", ">")
+        .replace("""\u0027""", "'")
+        .replace("""\/""", "/")
+        .replace("""\"""", """"""")
+        .replace("""\t""", "    ")
+        .replace("""\\""", """\""")
+        .replace("""\ """, "")
+        .trim()
 }
