+++ gradle.properties
@@ -28,8 +28,8 @@ sample_aps_banner=
 sample_aps_static=
 sample_aps_video=
 
-# Facebook Audience Network Ad Unit IDs
-# These are not required by Nimbus but enabled rendering of sample Facebook ads
+# Meta Audience Network Ad Unit IDs
+# These are not required by Nimbus but enabled rendering of sample Meta ads
 sample_fan_native_id=
 sample_fan_native_320_id=
 sample_fan_banner_320_id=
@@ -43,7 +43,3 @@ sample_unity_game_id=
 
 # Vungle Demand
 sample_vungle_config_id=
-sample_vungle_interstitial_id=
-sample_vungle_rewarded_id=
-sample_vungle_banner_320_id=
-sample_vungle_mrec_id=
